扣费统计  单价统计  BUG修复  SQL 语句少了 group by
git-svn-id: https://desktop-ee61sbk:8090/svn/remoteMeter/yccb/remoteMeter@147 4282536d-2d76-4d18-87bf-2ecf27506aab

+++ src/com/xdkj/yccb/main/charge/dto/SettleSum.java
@@ -1,11 +1,13 @@
 package com.xdkj.yccb.main.charge.dto;
 
+import java.math.BigDecimal;
+
 
 public class SettleSum {
 	
 	private String pricekindname;
 	private int yl;
-	private double demoney;
+	private BigDecimal demoney;
 	public String getPricekindname() {
 		return pricekindname;
 	}
@@ -18,10 +20,10 @@ public class SettleSum {
 	public void setYl(int yl) {
 		this.yl = yl;
 	}
-	public double getDemoney() {
+	public BigDecimal getDemoney() {
 		return demoney;
 	}
-	public void setDemoney(double demoney) {
+	public void setDemoney(BigDecimal demoney) {
 		this.demoney = demoney;
 	}
 	public SettleSum() {
